wip: snapshot of in-tree-fork patch attempt for FalkorDB embedding
Reference snapshot of the in-tree-fork patch we carried on top of
v8.6.0 while exploring how to embed RediSearch in FalkorDB. This
approach is being abandoned in favor of upstream LLAPI additions on
the falkordb/llapi-extensions-8.6 branch (see FalkorDB design doc
"temporal-prancing-tide"). Kept here so the prior code is greppable
while authoring the seven topical commits cleanly.

+++ src/redisearch_rs/trie_rs/src/node/accessors.rs
@@ -33,8 +33,12 @@ impl<Data> Node<Data> {
     }
 
     /// Returns the number of children for this node.
+    ///
+    /// FalkorDB patch: return type widened from `u8` to `u16` in lockstep with
+    /// `NodeHeader::n_children`. See the long comment on that field for the
+    /// rationale (256-child overflow on binary keys).
     #[inline]
-    pub const fn n_children(&self) -> u8 {
+    pub const fn n_children(&self) -> u16 {
         self.header().n_children
     }
